+++ i21Database/dbo/Stored Procedures/uspARCustomerStatementPaymentActivityReport.sql
@@ -13,6 +13,7 @@
 	, @strCustomerIds					AS NVARCHAR(MAX)	= NULL
 	, @ysnEmailOnly						AS BIT				= NULL
 	, @ysnIncludeWriteOffPayment    	AS BIT 				= 1
+	, @intEntityUserId					AS INT				= NULL
 AS
 
 SET QUOTED_IDENTIFIER OFF
@@ -40,6 +41,7 @@ DECLARE @dtmDateToLocal						AS DATETIME			= NULL
 	  , @queryBudget						AS NVARCHAR(MAX)
 	  , @filter								AS NVARCHAR(MAX)	= ''
 	  , @intWriteOffPaymentMethodId			AS INT				= NULL
+	  , @intEntityUserIdLocal				AS INT				= NULL
 
 DECLARE @temp_xml_table TABLE (
 	 [id]			INT IDENTITY(1,1)
@@ -120,6 +122,7 @@ SET @strCustomerNameLocal				= NULLIF(@strCustomerName, '')
 SET @strCustomerIdsLocal				= NULLIF(@strCustomerIds, '')
 SET @strDateTo							= ''''+ CONVERT(NVARCHAR(50),@dtmDateToLocal, 110) + ''''
 SET @strDateFrom						= ''''+ CONVERT(NVARCHAR(50),@dtmDateFromLocal, 110) + ''''
+SET @intEntityUserIdLocal				= NULLIF(@intEntityUserId, 0)
 
 IF @strCustomerNumberLocal IS NOT NULL
 	BEGIN
@@ -209,12 +212,13 @@ IF @ysnIncludeWriteOffPaymentLocal = 1
 		WHERE UPPER(strPaymentMethod) = 'WRITE OFF'
 	END
 
-TRUNCATE TABLE tblARCustomerAgingStagingTable
+DELETE FROM tblARCustomerAgingStagingTable WHERE intEntityUserId = @intEntityUserIdLocal AND strAgingType = 'Summary'
 INSERT INTO tblARCustomerAgingStagingTable (
 		  strCustomerName
 		, strCustomerNumber
 		, strCustomerInfo
 		, intEntityCustomerId
+		, intEntityUserId
 		, dblCreditLimit
 		, dblTotalAR
 		, dblFuture
@@ -234,10 +238,12 @@ INSERT INTO tblARCustomerAgingStagingTable (
 		, strSourceTransaction
 		, strCompanyName
 		, strCompanyAddress
+		, strAgingType
 )
 EXEC dbo.[uspARCustomerAgingAsOfDateReport] @strCompanyLocation = @strLocationNameLocal
 										  , @strCustomerName = @strCustomerNameLocal
 										  , @ysnIncludeWriteOffPayment = @ysnIncludeWriteOffPaymentLocal
+										  , @intEntityUserId = @intEntityUserIdLocal
 
 SET @query = CAST('' AS NVARCHAR(MAX)) + 
 'SELECT * 
@@ -482,7 +488,7 @@ IF @ysnPrintOnlyPastDueLocal = 1
 IF @ysnPrintZeroBalanceLocal = 0
     BEGIN
         DELETE FROM @temp_statement_table WHERE (((ABS(dblBalance) * 10000) - CONVERT(FLOAT, (ABS(dblBalance) * 10000))) <> 0) OR ISNULL(dblBalance, 0) = 0
-		DELETE FROM tblARCustomerAgingStagingTable WHERE (((ABS(dblTotalAR) * 10000) - CONVERT(FLOAT, (ABS(dblTotalAR) * 10000))) <> 0) OR ISNULL(dblTotalAR, 0) = 0
+		DELETE FROM tblARCustomerAgingStagingTable WHERE ((((ABS(dblTotalAR) * 10000) - CONVERT(FLOAT, (ABS(dblTotalAR) * 10000))) <> 0) OR ISNULL(dblTotalAR, 0) = 0) AND intEntityUserId = @intEntityUserIdLocal AND strAgingType = 'Summary'
     END
 
 IF @ysnPrintCreditBalanceLocal = 0
@@ -538,7 +544,7 @@ WHEN NOT MATCHED BY TARGET THEN
 INSERT (strEntityNo, dtmLastStatementDate, dblLastStatement)
 VALUES (strCustomerNumber, dtmLastStatementDate, dblLastStatement);
 
-TRUNCATE TABLE tblARCustomerStatementStagingTable
+DELETE FROM tblARCustomerStatementStagingTable WHERE intEntityUserId = @intEntityUserIdLocal AND strStatementFormat = 'Payment Activity'
 INSERT INTO tblARCustomerStatementStagingTable (
 	  intEntityCustomerId
 	, strCustomerNumber
@@ -576,7 +582,10 @@ INSERT INTO tblARCustomerStatementStagingTable (
 	, dblCredits
 	, dblPrepayments
 	, dtmAsOfDate
-	, blbLogo)
+	, blbLogo
+	, intEntityUserId
+	, strStatementFormat
+)
 SELECT MAINREPORT.*
 	 , dblCreditAvailable			= MAINREPORT.dblCreditLimit - ISNULL(AGINGREPORT.dblTotalAR, 0)
 	 , dblFuture					= ISNULL(AGINGREPORT.dblFuture, 0)
@@ -590,6 +599,8 @@ SELECT MAINREPORT.*
 	 , dblPrepayments				= ISNULL(AGINGREPORT.dblPrepayments, 0)
 	 , dtmAsOfDate					= @dtmDateToLocal
 	 , blbLogo						= dbo.fnSMGetCompanyLogo('Header')
+	 , intEntityUserId				= @intEntityUserIdLocal
+	 , strStatementFormat			= 'Payment Activity'
 FROM (
 	--- Without CF Report
 	SELECT intEntityCustomerId			= STATEMENTREPORT.intEntityCustomerId
@@ -659,8 +670,12 @@ FROM (
 ) MAINREPORT
 LEFT JOIN tblARCustomerAgingStagingTable AS AGINGREPORT 
 	ON MAINREPORT.intEntityCustomerId = AGINGREPORT.intEntityCustomerId
+	AND AGINGREPORT.intEntityUserId = @intEntityUserIdLocal
+	AND AGINGREPORT.strAgingType = 'Summary'
 INNER JOIN #CUSTOMERS CUSTOMER ON MAINREPORT.intEntityCustomerId = CUSTOMER.intEntityCustomerId
 
 
 UPDATE tblARCustomerStatementStagingTable
-	SET strComment = dbo.fnEMEntityMessage(intEntityCustomerId, 'Statement')
+SET strComment = dbo.fnEMEntityMessage(intEntityCustomerId, 'Statement')
+WHERE intEntityUserId = @intEntityUserIdLocal
+  AND strStatementFormat = 'Payment Activity'
